move iterator.rs to execution mod
